Commit with solution for day 1
Details:

In this commit, I work with Rust for the first time and attempt to solve
day one of the 2024 edition of advent of code.

After solving the challenge for day 1, I spent some time getting more
familiar with enumns as a data structure and using them to organise the
code a bit more.

I also decided to make each day it's own module within the project and
then use a handy `Day` enum as a convenient handler for executing the
code.

Given that each day of advent of code would always take the contents of
a file as input and return a string as output and also that there would
always be two parts this helped define the required interface for the
enum.

With this current implementation, subsequent solutions for other days
would not require editing the `main.rs` file and I can proceed with
adding a module for the new day and it's subsequent entry in the Day
enum.